backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
requests
pytesseract
pymupdf
Pillow
python-multipart
